refactor to match naming pattern
