+++ src/app/file/file-detail/file-detail.component.spec.ts
