+++ src/app/file/file-detail/file-detail.component.ts
