+++ src/app/file/file.component.html
